+++ R/mirt.R
@@ -286,7 +286,7 @@
 #'     stable for computing the information matrix. Default is .0001.
 #'     
 #'     If \code{SE.type = 'SEM'}, this is the tollerance used to terminate the S-EM computations for each parameter,
-#'     and if not specified the default is \code{sqrt(TOL)}}
+#'     and if not specified the default is \code{.001}}
 #'     \item{NCYCLES}{maximum number of EM or MH-RM cycles; defaults are 500 and 2000}
 #'     \item{BURNIN}{number of burn in cycles (stage 1) in MH-RM; default 150}
 #'     \item{SEMCYCLES}{number of SEM cycles (stage 2) in MH-RM; default 50}